Main: fix win burns
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@13850 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManage.cpp
@@ -2802,7 +2802,7 @@ static int handleList(int argc, char *argv[],
             CHECK_ERROR(hardDiskFormats [i],
                         COMGETTER(Id) (description.asOutParam()));
 
-            unsigned int caps;
+            ULONG caps;
             CHECK_ERROR(hardDiskFormats [i],
                         COMGETTER(Capabilities) (&caps));
